+++ Modules/Learn/Entities/HomeWorkSolution.php
@@ -23,6 +23,6 @@ class HomeWorkSolution extends Model
 
     public function attachs()
     {
-        return $this->morphOne(HomeWorkAttach::class, "owner");
+        return $this->morphMany(HomeWorkAttach::class, "owner");
     }
 }