+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Student User Api And Dashboard  .
 - Add Subject Module .
 - End Basic Crud for Teacher function  .
+- End Api for HomeWork End point .